chore: extract bank selection card to separate component SelectionCard
